refactor: sunset local storage for keyrings, update keyring + recovery UX
