+++ frontend/app/[team]/page.tsx
@@ -10,6 +10,7 @@ import { GetStarted } from '@/components/dashboard/GetStarted'
 import clsx from 'clsx'
 import { FaChevronRight } from 'react-icons/fa'
 import { Button } from '@/components/common/Button'
+import UnlockKeyringDialog from '@/components/auth/UnlockKeyringDialog'
 
 export default function AppsHome({ params }: { params: { team: string } }) {
   const { activeOrganisation: organisation } = useContext(organisationContext)
