+++ frontend/app/page.tsx
@@ -4,18 +4,15 @@ import { useRouter } from 'next/navigation'
 import { useContext, useEffect, useState } from 'react'
 import Loading from './loading'
 import { OrganisationType } from '@/apollo/graphql'
-import { getLocalKeyrings } from '@/utils/localStorage'
 import { organisationContext } from '@/contexts/organisationContext'
 import { Button } from '@/components/common/Button'
 import { FaArrowRight } from 'react-icons/fa'
-import { useSession } from 'next-auth/react'
 import { HeroPattern } from '@/components/common/HeroPattern'
 import { LogoWordMark } from '@/components/common/LogoWordMark'
 import { RoleLabel } from '@/components/users/RoleLabel'
 
 export default function Home() {
   const router = useRouter()
-  const { data: session } = useSession()
 
   const { organisations, activeOrganisation, setActiveOrganisation, loading } =
     useContext(organisationContext)
@@ -23,41 +20,17 @@ export default function Home() {
   const [showOrgCards, setShowOrgCards] = useState<boolean>(false)
 
   const handleRouteToOrg = (org: OrganisationType) => {
-    const localOrgs = getLocalKeyrings()
-
-    if (
-      localOrgs?.find(
-        (localOrg) => localOrg.org.id === org!.id && localOrg.email === session?.user?.email
-      )
-    ) {
-      router.push(`/${org!.name}`)
-    } else {
-      router.push(`${org!.name}/newdevice`)
-    }
+    router.push(`/${org!.name}`)
   }
 
   useEffect(() => {
     if (!loading && organisations !== null) {
-      const localOrgs = getLocalKeyrings()
-
       // if there is no org setup on the server, send to onboarding page
       if (organisations.length === 0) router.push('/signup')
       else if (organisations.length === 1) {
         const organisation = organisations[0]
         setActiveOrganisation(organisation)
-        // if local keyring exists on device for active organisation
-        if (
-          localOrgs?.find(
-            (localOrg) =>
-              localOrg.org.id === organisation!.id && localOrg.email === session?.user?.email
-          )
-        ) {
-          router.push(`/${organisation!.name}`)
-        }
-        // if no keyring on device, send to new device login page
-        else {
-          router.push(`${organisation!.name}/newdevice`)
-        }
+        router.push(`/${organisation!.name}`)
       } else {
         setShowOrgCards(true)
       }
